+++ src/osd/ReplicatedBackend.cc
@@ -1310,8 +1310,7 @@ void ReplicatedBackend::calc_head_subsets(
 	     << " overlap " << prev << dendl;
   }
 
-
-  if (cloning.num_intervals() > cct->_conf->osd_recover_clone_overlap_limit) {
+  if (static_cast<uint32_t>(cloning.num_intervals()) > cct->_conf->osd_recover_clone_overlap_limit) {
     dout(10) << "skipping clone, too many holes" << dendl;
     clone_subsets.clear();
     cloning.clear();
@@ -1395,7 +1394,7 @@ void ReplicatedBackend::calc_clone_subsets(
 	     << " overlap " << next << dendl;
   }
 
-  if (cloning.num_intervals() > cct->_conf->osd_recover_clone_overlap_limit) {
+  if (static_cast<uint32_t>(cloning.num_intervals()) > cct->_conf->osd_recover_clone_overlap_limit) {
     dout(10) << "skipping clone, too many holes" << dendl;
     clone_subsets.clear();
     cloning.clear();
